componentize app files
